Revert "Merge remote-tracking branch 'origin/parallax-implementation'" revert This reverts commit 51af06086cc9cc7eb5805f3c4e13110c8fa94cee, reversing changes made to 22ac19ab45ba2e74f9e247d5e8951feb20f8d743.

+++ routes/web.php
@@ -1,7 +1,6 @@
 <?php
 
 use App\Http\Controllers\AdminController;
-use App\Http\Controllers\BlogController;
 use App\Http\Controllers\ContactController;
 use App\Http\Controllers\GalleryController;
 use App\Http\Controllers\HomeController;
@@ -32,18 +31,12 @@ Route::fallback(function () {
 
 Route::get('/', [HomeController::class, 'index'])->name('home');
 Route::get('/gallery', [GalleryController::class, 'index'])->name('gallery');
-Route::get('/blog', [BlogController::class, 'index'])->name('blog');
-Route::post('/contact-us', [ContactController::class, 'store'])->name('contact.us.store');
-Route::get('/login', [SessionsController::class, 'login'])->name('login');
-Route::post('/sessions', [SessionsController::class, 'store'])->name('session.store');
+Route::post('contact-us', [ContactController::class, 'store'])->name('contact.us.store');
+Route::get('login', [SessionsController::class, 'login']);
+Route::post('sessions', [SessionsController::class, 'store']);
+Route::post('logout', [SessionsController::class, 'destroy']);
 
-Route::group(['middleware' => 'auth'], function () {
-    Route::middleware('can:admin')->group(function () {
-        Route::get('admin', [AdminController::class, 'index'])->name('admin');
-        Route::get('photos', [AdminController::class, 'photos'])->name('photos');
-        Route::post('logout', [SessionsController::class, 'destroy'])->withoutMiddleware(['auth']);
-        Route::get('/admin/illustrations', [AdminController::class, 'illustrations'])->name('illustrations');
-        Route::post('/admin/illustrations/store', [AdminController::class, 'store'])->name('illustrations.store');
-        Route::delete('admin/photos/{photo}', [AdminController::class, 'destroy']);
-    });
-});
+Route::get('admin', [AdminController::class, 'index'])->middleware('can:admin')->name('admin');
+Route::get('admin/photos/upload', [AdminController::class, 'create'])->middleware('can:admin');
+Route::post('admin/photos', [AdminController::class, 'store'])->middleware('can:admin');
+Route::delete('admin/photos/{photo}', [AdminController::class, 'destroy'])->middleware('can:admin');
